Change itemSelection_SetTracking() first argument from int to double for consistency with CharStyle::setTracking()
git-svn-id: svn://scribus.net/branches/Version16x/Scribus@26696 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scribusdoc.cpp
@@ -8987,7 +8987,7 @@ void ScribusDoc::itemSelection_resetOpticalMargins(Selection* customSelection)
 //	itemSelection_ApplyParagraphStyle(newStyle, customSelection);
 }
 
-void ScribusDoc::itemSelection_SetTracking(int kern, Selection* customSelection)
+void ScribusDoc::itemSelection_SetTracking(double kern, Selection* customSelection)
 {
 	CharStyle newStyle;
 	newStyle.setTracking(kern);
